# haltestellen introduce polygon geometry for tootip intersection