this is a postgis based approach to solve the issues differences between label extent and feature geometry extent. I'm not convinced that vector tiles will help anything here https://github.com/geoadmin/mf-geoadmin3/issues/3065.

[Testlink]( https://mf-geoadmin3.dev.bgdi.ch/ltclm/?topic=ech&lang=de&bgLayer=ch.swisstopo.pixelkarte-farbe&layers=ch.bav.haltestellen-oev&X=247821.90&Y=682731.97&zoom=12 )

The new polygon geometry is generated on the fly based on the attributes ``verkehrsmittel`` and ``betriebspunkttyp``

The new db view bav.oev_haltestellen_tooltip can be reviewed here:
https://github.com/geoadmin/db/commit/dd65846e57816fcd93a8733273fe0e989d9ec5fc