@davidoesch 2016-03-24
+1 tested on chrome FF and IE11: it works like a charm and mobile
-> perfect solution