@gjn 2016-03-24
Special treatement for each layer. It's maintenance heavy and will need attention for every layer on every update. It needs to be integrated in DI process.

@ltclm @AFoletti feasable?

Maybe we add the as temporary fix for this layer - and wait until we have a solution with vector tiles and then compare results and effort required for each layer to maintain it.