@ltclm 2016-03-24
there is only an "issue" when the extent of the label is not identical with the extent of the feature geometry. so manly point layers are affected, need some special treatement, there is no need to have this for all our layers.
I dont think that vector tiles would fix this particular problem.