@AFoletti 2016-03-24
should be updates-transparent since the new geometry is prepared by the view itself. It's a very nice solution for the present case and can help us handle borderline layers where point symbolization is big. Another candidate is luftfahrthindernis