@gjn 2016-03-24
I like the solution and the effect. Much improved.

I just want to avoid adding unnecessary complexity that increases our maintenance and testing tasks. Just adds another element to always consider. And views have their difficulty :=)

Speaking of tests: would be great to add some automated tests for this!

It's not only point layers, but also polyline layers, depending on how they are styled (I seem to remember similar troubles with IVS layers).

I've nothing against merging this, but I still hope that vector tiles would handle this better. @ltclm I wonder why you think vector tiles won't be able to handle this? Could you elaborate?