@gjn 2016-03-24
If I understand correctly, when we change the symbology, we need to adapt the view. Correct?