@ltclm 2016-03-24
yes if we change the icon size in mapserver we have to update the view definition too.

im not 100% sure how we will use vector tiles here.
if you will use the vector tiles to replace the source of the hover effect (hand), then i think it will improve the usability. because then we will use the original feature geomerty for hover. we will see... vector tiles will also introduce a lot of special treatements (symbology etc)...