@gjn 2016-03-24
Understood.

The idea with vector tiles is to use the symbol for detection - and each feature will directly have it's id as part of the transfered vector data. We don't even need an identify request to find out which elements are under a cursor. That's the theory.

Of course, vector tiles adds a whole lot of complexity on it's own.