@gjn 2016-03-29
@loicgasser Would like to have your opinion. To me, this could be merged for tomorrow.