@loicgasser 2016-03-29
not against that, but how does the identify request performs as compared to previousely?