@ltclm 2016-03-29
i have tried to address the comments of @loicgasser, thanks for your help!

the impact on the query performance is massive:

```sql
-- original query
"Limit  (cost=0.28..8.56 rows=1 width=100) (actual time=2.374..2.376 rows=1 loops=1)"
"  ->  Index Scan using oev_haltestellen_the_geom on oev_haltestellen  (cost=0.28..8.56 rows=1 width=100) (actual time=2.373..2.374 rows=1 loops=1)"
"        Index Cond: (the_geom && '0103000020155500000100000005000000000000004F4B2341000000007CD80441000000004F4B2341000000001CD9044100000000774B2341000000001CD9044100000000774B2341000000007CD80441000000004F4B2341000000007CD80441'::geometry)"
"        Filter: (('01010000201555000000000000634B234100000000CCD80441'::geometry && st_expand(the_geom, 10::double precision)) AND _st_dwithin(the_geom, '01010000201555000000000000634B234100000000CCD80441'::geometry, 10::double precision))"
"Planning time: 0.290 ms"
"Execution time: 2.451 ms"

-- new query
"Limit  (cost=0.00..279.68 rows=201 width=100) (actual time=65.710..255.636 rows=1 loops=1)"
"  ->  Seq Scan on oev_haltestellen  (cost=0.00..11982.90 rows=8612 width=100) (actual time=65.708..255.632 rows=1 loops=1)"
"        Filter: st_dwithin(CASE WHEN (betriebspunkttyp = 'Haltestelle'::text) THEN st_envelope(st_setsrid(st_makeline(st_makepoint((st_x(st_envelope(the_geom)) - ((18 * array_length(regexp_split_to_array(verkehrsmittel, '\_'::text), 1)))::double precision) (...)"
"        Rows Removed by Filter: 25835"
"Planning time: 0.372 ms"
"Execution time: 255.769 ms"
```

I will prepare a materializev view with the necessary geometry indices for a performant identify request.