@ltclm 2016-03-29
the matview has been created and is working fine:
```sql
"Limit  (cost=0.28..8.56 rows=1 width=164) (actual time=1.343..1.344 rows=1 loops=1)"
"  ->  Index Scan using oev_haltestellen_tooltip_bgdi_geom_poly on oev_haltestellen_tooltip  (cost=0.28..8.56 rows=1 width=164) (actual time=1.341..1.342 rows=1 loops=1)"
"        Index Cond: (bgdi_geom_poly && '010300002015550000010000000500000000000000EAD524410000000096400E4100000000EAD5244100000000E6400E4100000000FED5244100000000E6400E4100000000FED524410000000096400E4100000000EAD524410000000096400E41'::geometry)"
"        Filter: (('01010000201555000000000000F4D5244100000000BE400E41'::geometry && st_expand(bgdi_geom_poly, 5::double precision)) AND _st_dwithin(bgdi_geom_poly, '01010000201555000000000000F4D5244100000000BE400E41'::geometry, 5::double precision))"
"Planning time: 12.675 ms"
"Execution time: 1.428 ms"
```
uvek_master has been deployed to dev and int.
https://github.com/geoadmin/db/commit/3b35696b58e360fd445be505c98c91817c0bc0b0
So it is ready from my side.