@ltclm 2016-03-30
yes, these changes should not be in this pr, i have removed them.