@gjn 2016-03-30
Thanks alot. Another question:

How long does the update of the matview take? I'm asking because AFAIK, this is an auto-update layer.