@gjn 2016-03-30
@ltclm Thanks alot for this. It's a great user experience improvement!